+++ packages/drizzle/db.ts
@@ -10,7 +10,6 @@ if (!connectionString) {
 }
 
 export const sql = postgres(connectionString, {
-  prepare: false,
   max: 10,
   idle_timeout: 100,
   connect_timeout: 10,